[BOLT] Add ColdTrampoline pass
Add a pass that inserts trampolines for "never-taken" conditional
branches on x86. The idea is to replace a 6-byte long conditional branch
on a hot path with a 2-byte version short branch to a nearby island that
contains a trampoline to cold code. Trampolines could be reused. When
they are not, each trampolines adds one byte to the code size, but
instruction cache footprint on the hot path is reduced.